Update gstime.dart, remove toInt() from jdayVal
Before this, package was returning same position (noon trajectory) for every given time.

+++ lib/propagation/gstime.dart
@@ -51,7 +51,7 @@ double? gstime(dynamic yearOrDate,
     if (jdayVal == null) {
       return null;
     }
-    return gstimeInternal(jdayVal.toInt());
+    return gstimeInternal(jdayVal);
   }
   return gstimeInternal(yearOrDate);
 }
